requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
dj-static==0.0.6
Django==4.0.3
django-bootstrap4==22.1
django-stdimage==5.3.0
gunicorn==20.1.0
Pillow==9.1.0
soupsieve==2.3.1
sqlparse==0.4.2
static3==0.7.0
whitenoise==6.0.0
